+++ driver/factory.go
@@ -1,3 +1,6 @@
+// Copyright © 2022 Ory Corp
+// SPDX-License-Identifier: Apache-2.0
+
 package driver
 
 import (
@@ -18,7 +21,7 @@ func New(ctx context.Context, stdOutOrErr io.Writer, sl *servicelocatorx.Options
 	}
 
 	ctxter := sl.Contextualizer()
-	if err := r.Init(ctx, ctxter); err != nil {
+	if err := r.Init(ctx, ctxter, dOpts...); err != nil {
 		r.Logger().WithError(err).Error("Unable to initialize service registry.")
 		return nil, err
 	}