+++ test/e2e/cypress/helpers/httpbin.ts
@@ -1,3 +1,6 @@
+// Copyright © 2022 Ory Corp
+// SPDX-License-Identifier: Apache-2.0
+
 import * as oauth2 from "./oauth2"
 
 export function checkToken(