+++ test/e2e/cypress/helpers/react.ts
@@ -1,3 +1,6 @@
+// Copyright © 2022 Ory Corp
+// SPDX-License-Identifier: Apache-2.0
+
 import { SPA_URL } from "./index"
 
 export const routes = {