+++ test/e2e/cypress/integration/profiles/oidc-provider/registration.spec.ts
@@ -1,3 +1,6 @@
+// Copyright © 2022 Ory Corp
+// SPDX-License-Identifier: Apache-2.0
+
 import { gen } from "../../../helpers"
 import * as uuid from "uuid"
 import * as oauth2 from "../../../helpers/oauth2"