+++ text/id.go
@@ -1,3 +1,6 @@
+// Copyright © 2022 Ory Corp
+// SPDX-License-Identifier: Apache-2.0
+
 package text
 
 // This file MUST not have any imports to modules that are not in the standard library.